refactor(sparse,topology): lift cg_solve into deep_causality_sparse as public API
The matrix-free Conjugate Gradient solver was pub(crate) inside
deep_causality_topology, and its own doc-comment proposed lifting it into
deep_causality_sparse. Do that so the upcoming BRCD foundations (and any
other crate needing an SPD solve) can compose on it.

- Move cg_solve + CgFailure into deep_causality_sparse::solver::cg, make
  them public, and re-export from the crate root. The solver stays generic
  over T: RealField, matrix-free, and free of external numeric deps.
- Relocate the CG unit tests to tests/solver/cg_tests.rs, register them in
  tests/mod.rs and a new `solver` rust_test_suite in tests/BUILD.bazel. The
  one assertion that reached into the private `dot` helper now computes the
  norm inline.
- Point deep_causality_topology's Hodge decomposition at the public
  deep_causality_sparse::{cg_solve, CgFailure}; remove the crate-local copy.
  The gauge-fixing helper subtract_mean_in_place (Hodge-specific, not part of
  the solver API) stays in topology's utils/cg_solver.rs with an updated doc.
