+++ deep_causality_sparse/src/solver/mod.rs
@@ -2,7 +2,7 @@
  * SPDX-License-Identifier: MIT
  * Copyright (c) 2023 - 2026. The DeepCausality Authors and Contributors. All Rights Reserved.
  */
-mod errors;
-mod extensions;
-mod solver;
-mod types;
+
+//! Iterative solvers over sparse / matrix-free operators.
+
+pub mod cg;
