+++ deep_causality_sparse/tests/BUILD.bazel
@@ -33,6 +33,21 @@ rust_test_suite(
 )
 
 rust_test_suite(
+    name = "solver",
+    srcs = glob([
+        "solver/*_tests.rs",
+    ]),
+    tags = [
+        "unit-test",
+    ],
+    visibility = ["//visibility:public"],
+    deps = [
+        # Crate to test
+        "//deep_causality_sparse",
+    ],
+)
+
+rust_test_suite(
     name = "extensions",
     srcs = glob([
         "extensions/*_tests.rs",
