+++ deep_causality_sparse/tests/solver/mod.rs
@@ -2,7 +2,5 @@
  * SPDX-License-Identifier: MIT
  * Copyright (c) 2023 - 2026. The DeepCausality Authors and Contributors. All Rights Reserved.
  */
-mod errors;
-mod extensions;
-mod solver;
-mod types;
+#[cfg(test)]
+mod cg_tests;
